Move files to aboutUs folder; remove unused variables
