+++ src/js/components/App.js
@@ -21,7 +21,7 @@ import TimetableContainer from "../containers/timetable/TimetableContainer";
 import Syllabus from "./syllabus/Syllabus";
 import CourseInfo from "./courseInfo/CourseInfo";
 import RedirectPage from "./user/RedirectPage";
-import AboutUs from "./AboutUs"
+import AboutUs from "./aboutUs/AboutUs";
 // import FooterContainer from "../containers/FooterContainer";
 import Bus from "./Bus";
 import Career from "./career/Career";
