+++ src/js/components/aboutUs/AboutUs.js
@@ -2,13 +2,12 @@ import React from "react";
 import styled from "styled-components";
 import { Tab, Menu } from "semantic-ui-react";
 import { withNamespaces } from "react-i18next";
-import Header from "./Header";
-import { Wrapper } from "../styled-components/Wrapper";
+import Header from "../Header";
+import { Wrapper } from "../../styled-components/Wrapper";
 import { Helmet } from "react-helmet";
-import { media } from "../styled-components/utils";
-import { Overlay } from "../styled-components/Overlay";
+import { media } from "../../styled-components/utils";
 
-import OurMission from "./OurMission.js";
+// import OurMission from "./OurMission.js";
 import MeetOurTeam from "./MeetOurTeam.js";
 import JoinUs from "./JoinUs.js";
 
@@ -21,17 +20,13 @@ const PlaceHolder = styled("div")`
 //   padding: 65px 0px 0px 0px;
 //   ${media.tablet`padding-top: 0em;`}
 // `;
-// 想要override semantic的style需要用「!important」
-// 內容的字變大了，但標籤的字沒受到影響？！
+
 const LargeTab = styled(Tab)`
   // text-align: center;
   // width: 80%
   margin: 0 4% !important;
   // border: solid;
 `;
-{
-  /* <img></img> */
-}
 
 const panes = (t) => [
   {
@@ -42,16 +37,14 @@ const panes = (t) => [
     ),
     render: () => <JoinUs />,
   },
-  {
-    // 要直接override menuItem本身的style（個案）
-    menuItem: (
-      <Menu.Item style={{ fontSize: "1.5em" }}>
-        {t("aboutus.our mission")}
-      </Menu.Item>
-    ),
-    render: () => <OurMission />,
-  },
-  // 其他的就麻煩你修改然後繼續開發～
+  // {
+  //   menuItem: (
+  //     <Menu.Item style={{ fontSize: "1.5em" }}>
+  //       {t("aboutus.our mission")}
+  //     </Menu.Item>
+  //   ),
+  //   render: () => <OurMission />,
+  // },
   {
     menuItem: (
       <Menu.Item style={{ fontSize: "1.5em" }}>
@@ -61,7 +54,6 @@ const panes = (t) => [
     render: () => <MeetOurTeam />,
   },
 ];
-// const { t, lng } = this.props;
 const AboutUs = ({ t }) => (
   <Wrapper>
     <Helmet>
